IO refactor WIP
