+++ LanguageExt.Core/Effects/IO/DSL/IOLiftAsync.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Threading.Tasks;
+
+namespace LanguageExt.DSL;
+    
+record IOLiftAsync<A>(Func<EnvIO, Task<A>> F) : IODsl<A>
+{
+    public override IODsl<B> Map<B>(Func<A, B> f) =>
+        new IOLiftAsync<B>(async x => f(await F(x).ConfigureAwait(false)));
+}
