+++ LanguageExt.Core/Effects/IO/ORIGINAL/IO.ORIGINAL.cs
@@ -1,10 +1,10 @@
+/*
 using System;
 using System.Diagnostics.Contracts;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using LanguageExt.Common;
-using LanguageExt.DSL;
 using LanguageExt.Traits;
 
 namespace LanguageExt;
@@ -27,8 +27,7 @@ namespace LanguageExt;
 /// </summary>
 /// <param name="runIO">The lifted thunk that is the IO operation</param>
 /// <typeparam name="A">Bound value</typeparam>
-public abstract record IO<A> :
-    K<IO, A>,
+public abstract record IO<A> : 
     Fallible<IO<A>, IO, Error, A>,
     Monoid<IO<A>>
 {
@@ -51,25 +50,36 @@ public abstract record IO<A> :
     public static IO<A> Empty { get; } = 
         new IOFail<A>(Errors.None);
 
+    internal abstract bool IsAsync { get; }
+
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     //
     //  Lifting
     //
-
-    static IO<A> Lift(IODsl<A> fa) => 
-        new IOBind<A>(fa.Map(IO.pure));
-    
-    public static IO<A> Lift(Func<EnvIO, A> f) => 
-        Lift(IODsl.Lift(f));
-    
-    public static IO<A> LiftAsync(Func<EnvIO, Task<A>> f) => 
-        Lift(IODsl.Lift(f));
     
     public static IO<A> Lift(Func<A> f) =>
-        Lift(IODsl.Lift(_ => f()));
+        new IOSync<A>(_ => IOResponse.Complete(f()));
+
+    public static IO<A> Lift(Func<IOResponse<A>> f) =>
+        new IOSync<A>(_ => f());
+
+    public static IO<A> Lift(Func<EnvIO, A> f) =>
+        new IOSync<A>(e => IOResponse.Complete(f(e)));
+
+    public static IO<A> Lift(Func<EnvIO, IOResponse<A>> f) =>
+        new IOSync<A>(f);
+
+    public static IO<A> LiftAsync(Func<Task<IOResponse<A>>> f) =>
+        new IOAsync<A>(async _ => await f());
+
+    public static IO<A> LiftAsync(Func<EnvIO, Task<IOResponse<A>>> f) =>
+        new IOAsync<A>(f);
 
     public static IO<A> LiftAsync(Func<Task<A>> f) =>
-        Lift(IODsl.Lift(_ => f()));
+        new IOAsync<A>(async _ => IOResponse.Complete(await f()));
+
+    public static IO<A> LiftAsync(Func<EnvIO, Task<A>> f) =>
+        new IOAsync<A>(async env => IOResponse.Complete(await f(env)));
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     //
@@ -78,8 +88,6 @@ public abstract record IO<A> :
 
     public abstract IO<B> Map<B>(Func<A, B> f);
 
-    public abstract IO<B> ApplyBack<B>(K<IO, Func<A, B>> f);
-    
     public IO<B> Map<B>(B value) =>
         Map(_ => value);
 
@@ -191,28 +199,11 @@ public abstract record IO<A> :
         Func<(S State, A Value), bool> predicate) =>
         FoldUntil(Schedule.Forever, initialState, folder, predicate);
 
-    public IO<S> FoldUntil<S>(
+    public abstract IO<S> FoldUntil<S>(
         Schedule schedule,
         S initialState,
         Func<S, A, S> folder,
-        Func<(S State, A Value), bool> predicate) =>
-        IO<S>.LiftAsync(async envIO =>
-                        {
-                            if (envIO.Token.IsCancellationRequested) throw new TaskCanceledException();
-                            var r     = await RunAsync(envIO);
-                            var state = folder(initialState, r);
-                            if (predicate((state, r))) return state;
-
-                            var token = envIO.Token;
-                            foreach (var delay in schedule.Run())
-                            {
-                                await IO.yieldFor(delay, token);
-                                r = await RunAsync(envIO);
-                                state = folder(state, r);
-                                if (predicate((state, r))) return state;
-                            }
-                            return state;
-                        });
+        Func<(S State, A Value), bool> predicate);
     
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     //
@@ -224,48 +215,17 @@ public abstract record IO<A> :
     /// of the IO chain (i.e. the one embedded within the `EnvIO` environment that is passed through
     /// all IO computations)
     /// </summary>
-    public IO<A> Post() =>
-        LiftAsync(async env =>
-                       {
-                           if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-                           if (env.SyncContext is null) return await RunAsync(env);
-
-                           A?         value = default;
-                           Exception? error = default;
-                           using var  wait  = new AutoResetEvent(false);
-
-                           env.SyncContext.Post(_ =>
-                                                {
-                                                    try
-                                                    {
-                                                        value = Run(env);
-                                                    }
-                                                    catch (Exception e)
-                                                    {
-                                                        error = e;
-                                                    }
-                                                    finally
-                                                    {
-                                                        wait.Set();
-                                                    }
-                                                },
-                                                null);
-
-                           await wait.WaitOneAsync(env.Token);
-                           if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-                           if (error is not null) error.Rethrow<A>();
-                           return value!;
-                       });
+    public abstract IO<A> Post();
     
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     //
     //  Monad
     //
 
-    public abstract IO<B> Bind<B>(Func<A, K<IO, B>> f);
+    public abstract IO<B> Bind<B>(Func<A, IO<B>> f);
 
-    public IO<B> Bind<B>(Func<A, IO<B>> f) =>
-        Bind(x => f(x).Kind());
+    public IO<B> Bind<B>(Func<A, K<IO, B>> f) =>
+        Bind(x => f(x).As());
 
     public IO<B> Bind<B>(Func<A, Pure<B>> f) =>
         Map(x => f(x).Value);
@@ -403,12 +363,7 @@ public abstract record IO<A> :
     /// </summary>
     [Pure]
     [MethodImpl(Opt.Default)]
-    public IO<A> Bracket() =>
-        LiftAsync(async env =>
-                  {
-                      using var lenv = env.LocalResources;
-                      return await RunAsync(lenv);
-                  });
+    public abstract IO<A> Bracket();
 
     /// <summary>
     /// When acquiring, using, and releasing various resources, it can be quite convenient to write a function to manage
@@ -428,23 +383,7 @@ public abstract record IO<A> :
     /// <param name="Use">Function to use the acquired resource</param>
     /// <param name="Catch">Function to run to handle any exceptions</param>
     /// <param name="Fin">Function to invoke to release the resource</param>
-    public IO<C> Bracket<B, C>(Func<A, IO<C>> Use, Func<Error, IO<C>> Catch, Func<A, IO<B>> Fin) =>
-        IO<C>.LiftAsync(async env =>
-                        {
-                            var x = await RunAsync(env);
-                            try
-                            {
-                                return await Use(x).RunAsync(env);
-                            }
-                            catch (Exception e)
-                            {
-                                return await Catch(e).RunAsync(env);
-                            }
-                            finally
-                            {
-                                (await Fin(x).RunAsync(env))?.Ignore();
-                            }
-                        });    
+    public abstract IO<C> Bracket<B, C>(Func<A, IO<C>> Use, Func<Error, IO<C>> Catch, Func<A, IO<B>> Fin);
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     //
@@ -455,7 +394,7 @@ public abstract record IO<A> :
         Pure(ma.Value);
 
     public static implicit operator IO<A>(Error error) =>
-        Lift(_ => error.Throw<A>());
+        Lift(_ => error.Throw<IOResponse<A>>());
 
     public static implicit operator IO<A>(Fail<Error> ma) =>
         Lift(() => ma.Value.Throw<A>());
@@ -485,21 +424,7 @@ public abstract record IO<A> :
     /// <summary>
     /// Create a local cancellation environment
     /// </summary>
-    public IO<A> Local() =>
-        LiftAsync(async env =>
-                  {
-                      if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-
-                      // Create a new local token-source with its own cancellation token
-                      using var tsrc = new CancellationTokenSource();
-                      var       tok  = tsrc.Token;
-
-                      // If the parent cancels, we should too
-                      await using var reg = env.Token.Register(() => tsrc.Cancel());
-
-                      var env1 = EnvIO.New(env.Resources, tok, tsrc, env.SyncContext);
-                      return await RunAsync(env1);
-                  });    
+    public abstract IO<A> Local();    
     
     /// <summary>
     /// Queues the specified work to run on the thread pool  
@@ -512,56 +437,44 @@ public abstract record IO<A> :
     /// </remarks>
     /// <param name="timeout">Optional timeout</param>
     /// <returns>`Fork` record that contains members for cancellation and optional awaiting</returns>
-    public IO<ForkIO<A>> Fork(Option<TimeSpan> timeout = default) =>
-        IO<ForkIO<A>>.Lift(
-            env =>
-            {
-                if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-                
-                // Create a new local token-source with its own cancellation token
-                var tsrc  = timeout.Match(Some: to => new CancellationTokenSource(to), 
-                                          None: () => new CancellationTokenSource());
-                var token = tsrc.Token;
+    public abstract IO<ForkIO<A>> Fork(Option<TimeSpan> timeout = default);
 
-                // If the parent cancels, we should too
-                var reg = env.Token.Register(() => tsrc.Cancel());
+    /// <summary>
+    /// Run the `IO` monad to get its result
+    /// </summary>
+    /// <remarks>
+    /// This forks the operation to run on a new task that is then awaited.
+    /// 
+    /// Any lifted asynchronous operations will yield to the thread-scheduler, allowing other queued
+    /// operations to run concurrently.  So, even though this call isn't awaitable it still plays
+    /// nicely and doesn't block the thread.
+    /// </remarks>
+    /// <remarks>
+    /// NOTE: An exception will always be thrown if the IO operation fails.  Lift this monad into
+    /// other error handling monads to leverage more declarative error handling. 
+    /// </remarks>
+    /// <returns>Result of the IO operation</returns>
+    /// <exception cref="TaskCanceledException">Throws if the operation is cancelled</exception>
+    /// <exception cref="BottomException">Throws if any lifted task fails without a value `Exception` value.</exception>
+    public abstract ValueTask<A> RunAsync(EnvIO? envIO = null);
 
-                // Gather our resources for clean-up
-                var cleanup = new CleanUp(tsrc, reg);
-
-                var parentResources = env.Resources;
-
-                var task = Task.Factory.StartNew(
-                    () =>
-                    {
-                        var forkedResources = new Resources(parentResources);
-                        try
-                        {
-                            var t = RunAsync(EnvIO.New(forkedResources, token, tsrc, env.SyncContext));
-                            return t.GetAwaiter().GetResult();
-                        }
-                        finally
-                        {
-                            forkedResources.Dispose();
-                            cleanup.Dispose();
-                        }
-                    }, TaskCreationOptions.LongRunning);
-
-                return new ForkIO<A>(
-                        IO<Unit>.Lift(() => {
-                                          try
-                                          {
-                                              tsrc.Cancel();
-                                          }
-                                          catch(ObjectDisposedException)
-                                          {
-                                              // ignore if already cancelled
-                                          }
-                                          return default; 
-                                      }),
-                        LiftAsync(e => AwaitAsync(task, e, token, tsrc)));
-            });
-    
+    /// <summary>
+    /// Run the `IO` monad to get its result
+    /// </summary>
+    /// <remarks>
+    /// Any lifted asynchronous operations will yield to the thread-scheduler, allowing other queued
+    /// operations to run concurrently.  So, even though this call isn't awaitable it still plays
+    /// nicely and doesn't block the thread.
+    /// </remarks>
+    /// <remarks>
+    /// NOTE: An exception will always be thrown if the IO operation fails.  Lift this monad into
+    /// other error handling monads to leverage more declarative error handling. 
+    /// </remarks>
+    /// <param name="env">IO environment</param>
+    /// <returns>Result of the IO operation</returns>
+    /// <exception cref="TaskCanceledException">Throws if the operation is cancelled</exception>
+    /// <exception cref="BottomException">Throws if any lifted task fails without a value `Exception` value.</exception>
+    public abstract A Run(EnvIO? envIO = null);
     
     /// <summary>
     /// Run the `IO` monad to get its result.  Differs from `Run` in that it catches any exceptions and turns
@@ -634,31 +547,8 @@ public abstract record IO<A> :
     /// </remarks>
     /// <param name="predicate">Keep repeating until this predicate returns `true` for each computed value</param>
     /// <returns>The result of the last invocation</returns>
-    public IO<A> RepeatUntil(Func<A, bool> predicate) =>
-        LiftAsync(async env =>
-                  {
-                      if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-                      var lenv = env.LocalResources;
-                      try
-                      {
-                          while (!env.Token.IsCancellationRequested)
-                          {
-                              var result = await RunAsync(lenv);
-
-                              // free any resources acquired during a repeat
-                              await lenv.Resources.ReleaseAll().RunAsync(env);
-
-                              if (predicate(result)) return result;
-                          }
-
-                          throw new TaskCanceledException();
-                      }
-                      finally
-                      {
-                          // free any resources acquired during a repeat
-                          await lenv.Resources.ReleaseAll().RunAsync(env);
-                      }
-                  });
+    public abstract IO<A> RepeatUntil(
+        Func<A, bool> predicate);
 
     /// <summary>
     /// Keeps repeating the computation, until the scheduler expires, or the predicate returns true, or an error occurs
@@ -670,42 +560,9 @@ public abstract record IO<A> :
     /// <param name="schedule">Scheduler strategy for repeating</param>
     /// <param name="predicate">Keep repeating until this predicate returns `true` for each computed value</param>
     /// <returns>The result of the last invocation</returns>
-    public IO<A> RepeatUntil(
+    public abstract IO<A> RepeatUntil(
         Schedule schedule,
-        Func<A, bool> predicate) =>
-        LiftAsync(async env =>
-                  {
-                      if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-                      var token = env.Token;
-                      var lenv  = env.LocalResources;
-                      try
-                      {
-                          var result = await RunAsync(lenv);
-
-                          // free any resources acquired during a repeat
-                          await lenv.Resources.ReleaseAll().RunAsync(env);
-
-                          if (predicate(result)) return result;
-
-                          foreach (var delay in schedule.Run())
-                          {
-                              await Task.Delay((TimeSpan)delay, token);
-                              result = await RunAsync(lenv);
-
-                              // free any resources acquired during a repeat
-                              await lenv.Resources.ReleaseAll().RunAsync(env);
-
-                              if (predicate(result)) return result;
-                          }
-
-                          return result;
-                      }
-                      finally
-                      {
-                          // free any resources acquired during a repeat
-                          await lenv.Resources.ReleaseAll().RunAsync(env);
-                      }
-                  });      
+        Func<A, bool> predicate);
     
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     //
@@ -784,34 +641,8 @@ public abstract record IO<A> :
     /// So, successive retries will not grow the acquired resources on each retry iteration.  Any successful operation that
     /// acquires resources will have them tracked in the usual way. 
     /// </remarks>
-    public IO<A> RetryUntil(Func<Error, bool> predicate) =>
-        LiftAsync(async env =>
-                  {
-                      if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-                      var lenv = env.LocalResources;
-
-                      while (!env.Token.IsCancellationRequested)
-                      {
-                          try
-                          {
-                              var r = await RunAsync(lenv);
-
-                              // Any resources that were acquired should be propagated through to the `env`
-                              env.Resources.Merge(lenv.Resources);
-
-                              return r;
-                          }
-                          catch (Exception e)
-                          {
-                              // Any resources created whilst trying should be removed for the retry
-                              await lenv.Resources.ReleaseAll().RunAsync(env);
-
-                              if (predicate(Error.New(e))) throw;
-                          }
-                      }
-
-                      throw new TaskCanceledException();
-                  });
+    public abstract IO<A> RetryUntil(
+        Func<Error, bool> predicate);
 
     /// <summary>
     /// Retry if the IO computation fails 
@@ -825,55 +656,9 @@ public abstract record IO<A> :
     /// So, successive retries will not grow the acquired resources on each retry iteration.  Any successful operation that
     /// acquires resources will have them tracked in the usual way. 
     /// </remarks>
-    public IO<A> RetryUntil(Schedule schedule, Func<Error, bool> predicate) =>
-        LiftAsync(async env =>
-                  {
-                      if (env.Token.IsCancellationRequested) throw new TaskCanceledException();
-                      var        token = env.Token;
-                      Exception? lastError;
-                      var        lenv = env.LocalResources;
-                      try
-                      {
-                          var r = await RunAsync(lenv);
-
-                          // Any resources that were acquired should be propagated through to the `env`
-                          env.Resources.Merge(lenv.Resources);
-
-                          return r;
-                      }
-                      catch (Exception e)
-                      {
-                          // Any resources created whilst trying should be removed for the retry
-                          await lenv.Resources.ReleaseAll().RunAsync(env);
-
-                          if (predicate(Error.New(e))) throw;
-                          lastError = e;
-                      }
-
-                      foreach (var delay in schedule.Run())
-                      {
-                          await Task.Delay((TimeSpan)delay, token);
-                          try
-                          {
-                              var r = await RunAsync(lenv);
-
-                              // Any resources that were acquired should be propagated through to the `env`
-                              env.Resources.Merge(lenv.Resources);
-
-                              return r;
-                          }
-                          catch (Exception e)
-                          {
-                              // Any resources created whilst trying should be removed for the retry
-                              await lenv.Resources.ReleaseAll().RunAsync(env);
-
-                              if (predicate(Error.New(e))) throw;
-                              lastError = e;
-                          }
-                      }
-
-                      return lastError.Rethrow<A>();
-                  });        
+    public abstract IO<A> RetryUntil(
+        Schedule schedule,
+        Func<Error, bool> predicate);
 
     /// <summary>
     /// Catches any error thrown by invoking this IO computation, passes it through a predicate,
@@ -882,8 +667,7 @@ public abstract record IO<A> :
     /// </summary>
     /// <param name="Predicate">Predicate</param>
     /// <param name="Fail">Fail functions</param>
-    public IO<A> Catch(Func<Error, bool> Predicate, Func<Error, K<IO, A>> Fail) =>
-        new IOCatch<A, A>(this, Predicate, Fail, Pure);
+    public abstract IO<A> Catch(Func<Error, bool> Predicate, Func<Error, K<IO, A>> Fail);
 
     /// <summary>
     /// Monoid combine
@@ -892,122 +676,8 @@ public abstract record IO<A> :
     /// <returns>This if computation runs without error.  `rhs` otherwise</returns>
     public IO<A> Combine(IO<A> rhs) =>
         Catch(_ => true, _ => rhs);
-
-    /// <summary>
-    /// Run the `IO` monad to get its result
-    /// </summary>
-    /// <remarks>
-    /// Any lifted asynchronous operations will yield to the thread-scheduler, allowing other queued
-    /// operations to run concurrently.  So, even though this call isn't awaitable it still plays
-    /// nicely and doesn't block the thread.
-    /// </remarks>
-    /// <remarks>
-    /// NOTE: An exception will always be thrown if the IO operation fails.  Lift this monad into
-    /// other error handling monads to leverage more declarative error handling. 
-    /// </remarks>
-    /// <param name="env">IO environment</param>
-    /// <returns>Result of the IO operation</returns>
-    /// <exception cref="TaskCanceledException">Throws if the operation is cancelled</exception>
-    /// <exception cref="BottomException">Throws if any lifted task fails without a value `Exception` value.</exception>
-    public A Run(EnvIO? envIO = null) =>
-        RunAsync(envIO).GetAwaiter().GetResult();
-
-    /// <summary>
-    /// Run the `IO` monad to get its result
-    /// </summary>
-    /// <remarks>
-    /// This forks the operation to run on a new task that is then awaited.
-    /// 
-    /// Any lifted asynchronous operations will yield to the thread-scheduler, allowing other queued
-    /// operations to run concurrently.  So, even though this call isn't awaitable it still plays
-    /// nicely and doesn't block the thread.
-    /// </remarks>
-    /// <remarks>
-    /// NOTE: An exception will always be thrown if the IO operation fails.  Lift this monad into
-    /// other error handling monads to leverage more declarative error handling. 
-    /// </remarks>
-    /// <returns>Result of the IO operation</returns>
-    /// <exception cref="TaskCanceledException">Throws if the operation is cancelled</exception>
-    /// <exception cref="BottomException">Throws if any lifted task fails without a value `Exception` value.</exception>
-    public async ValueTask<A> RunAsync(EnvIO? envIO = null)
-    {
-        if (envIO?.Token.IsCancellationRequested ?? false) throw new TaskCanceledException();
-        var envRequiresDisposal = envIO is null;
-        envIO ??= EnvIO.New();
-        var ma = this;
-
-        try
-        {
-            while (!envIO.Token.IsCancellationRequested)
-            {
-                switch (ma)
-                {
-                    case IOPure<A> (var value):
-                        return value;
-
-                    case IOPureAsync<A> (var value):
-                        return await value;
-                            
-                    case IOFail<A> (var value):
-                        return value.Throw<A>();
-
-                    case IOCatch<A> @catch:
-                        ma = await @catch.Invoke(envIO);
-                        break;
-
-                    case IOBind<A> (var dsl):
-                        switch (dsl)
-                        {
-                            case IOLiftSync<IO<A>> (var f):
-                                ma = f(envIO);
-                                break;
-
-                            case IOLiftAsync<IO<A>> (var f):
-                                ma = await f(envIO);
-                                break;
-
-                            default:
-                                throw new InvalidOperationException("We shouldn't be here!");
-                        }
-                        break;
-                }
-            }
-
-            throw new TaskCanceledException();
-        }
-        finally
-        {
-            if (envRequiresDisposal) envIO.Dispose();
-        }
-    }    
     
     public override string ToString() => 
         "IO";
-    
-    
-    ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    //
-    //  Internal
-    //
-    
-    async Task<A> AwaitAsync(Task<A> t, EnvIO envIO, CancellationToken token, CancellationTokenSource source)
-    {
-        if (envIO.Token.IsCancellationRequested) throw new TaskCanceledException();
-        if (token.IsCancellationRequested) throw new TaskCanceledException();
-        await using var reg = envIO.Token.Register(source.Cancel);
-        return await t;        
-    }
-    
-    record CleanUp(CancellationTokenSource Src, CancellationTokenRegistration Reg) : IDisposable
-    {
-        volatile int disposed;
-        public void Dispose()
-        {
-            if (Interlocked.Exchange(ref disposed, 1) == 0)
-            {
-                try{Src.Dispose();} catch { /* not important */ }
-                try{Reg.Dispose();} catch { /* not important */ }
-            }
-        }
-    }
 }
+*/
