+++ LanguageExt.Core/Effects/IO/ORIGINAL/Pure/IO.cs
@@ -1,3 +1,4 @@
+/*
 using System;
 using System.Diagnostics.Contracts;
 using System.Runtime.CompilerServices;
@@ -26,56 +27,60 @@ namespace LanguageExt;
 /// </summary>
 /// <param name="runIO">The lifted thunk that is the IO operation</param>
 /// <typeparam name="A">Bound value</typeparam>
-record IOFail<A>(Error Error) : IO<A>
+record IOPure<A>(A Value) : IO<A>
 {
     internal override bool IsAsync =>
         false;
-    
+
     public IO<A> ToSync() =>
-        new IOSync<A>(_ => throw Error.ToErrorException());
+        new IOSync<A>(_ => IOResponse.Complete(Value));
     
     public IO<A> ToAsync() =>
-        new IOAsync<A>(_ => throw Error.ToErrorException());
-    
+        new IOAsync<A>(_ => Task.FromResult(IOResponse.Complete(Value)));
+
     public override IO<B> Map<B>(Func<A, B> f) =>
-        new IOFail<B>(Error);
+        new IOSync<B>(_ => IOResponse.Complete(f(Value)));
 
     public override IO<S> FoldUntil<S>(
         Schedule schedule, 
         S initialState, 
         Func<S, A, S> folder,
         Func<(S State, A Value), bool> predicate) =>
-        new IOFail<S>(Error);
+        ToAsync().FoldUntil(schedule, initialState, folder, predicate);
 
     public override IO<A> Post() => 
         this;
 
-    public override IO<B> Bind<B>(Func<A, IO<B>> f) => 
-        new IOFail<B>(Error);
+    public override IO<B> Bind<B>(Func<A, IO<B>> f) =>
+        f(Value);
 
     public override IO<A> Bracket() => 
         this;
 
     public override IO<C> Bracket<B, C>(Func<A, IO<C>> Use, Func<Error, IO<C>> Catch, Func<A, IO<B>> Finally) => 
-        new IOFail<C>(Error);
+        ToSync().Bracket(Use, Catch, Finally);
 
     public override IO<A> Local() => 
         this;
 
     public override IO<ForkIO<A>> Fork(Option<TimeSpan> timeout = default) => 
-        new IOFail<ForkIO<A>>(Error);
+        new IOPure<ForkIO<A>>(new ForkIO<A>(Prelude.unitIO, this));
 
     public override ValueTask<A> RunAsync(EnvIO? envIO = null) => 
-        ValueTask.FromException<A>(Error.ToErrorException());
+        ValueTask.FromResult(Value);
 
     public override A Run(EnvIO? envIO = null) => 
-        Error.ToErrorException().Rethrow<A>();
+        Value;
 
     public override IO<A> RepeatUntil(Func<A, bool> predicate) =>
-        this;
+        predicate(Value)
+            ? this
+            : throw new InvalidOperationException("non-terminating and non-effectful IO repeat");
 
     public override IO<A> RepeatUntil(Schedule schedule, Func<A, bool> predicate) => 
-        this;
+        predicate(Value)
+            ? this
+            : throw new InvalidOperationException("non-terminating and non-effectful IO repeat");
 
     public override IO<A> RetryUntil(Func<Error, bool> predicate) => 
         this;
@@ -91,7 +96,6 @@ record IOFail<A>(Error Error) : IO<A>
     /// <param name="Predicate">Predicate</param>
     /// <param name="Fail">Fail functions</param>
     public override IO<A> Catch(Func<Error, bool> Predicate, Func<Error, K<IO, A>> Fail) =>
-        Predicate(Error)
-            ? Fail(Error).As()
-            : this;
+        this;
 }
+*/
